+++ src/sas/qtgui/Perspectives/Corfunc/CorfunSlider.py
