+++ src/sas/qtgui/Perspectives/Corfunc/util.py
